requirements.txt
torch>=1.13.1
torchvision>=0.14.1
torchaudio>=0.13.1
pytorch-lightning>=2.0.0
numpy>=1.22.4
scipy>=1.9.3
matplotlib>=3.6.2
transformers>=4.28.1
datasets>=2.10.1
Pillow>=9.3.0
sentencepiece
tqdm>=4.64.1
h5py>=3.7.0
faiss-gpu>=1.7.2
scikit-learn>=1.2.1
jupyter>=1.0.0
stream-chat>=4.0.0
streamlit>=1.18.1
gradio>=3.19.1